# jwt: Parse standard claims faster